# Conflicts:
#	app/src/main/java/com/example/dattingapp/Activity/Discovery.java

+++ app/src/main/java/com/example/dattingapp/DTO/DiscorverRequest.java
@@ -5,5 +5,5 @@ public class DiscorverRequest {
     public  int minAge;
     public  int maxAge;
     public String gender;
-    public Double distance;
+    public int distance;
 }
